Add configuration management and tests
A new `Config` class has been added to handle reading and writing to a configuration file. This class is used in the main application to prompt the user for configuration details. The `faker` package has been added as a dependency for testing purposes. New tests have been added to verify the functionality of the `Config` class.